Mudando apenas o caminho do import, pois os componentes estavam todos dentro da pasta UI
