+++ src/components/About/AboutText.tsx
@@ -1,4 +1,4 @@
-import { useLanguage } from '../../../hooks/languageHook'
+import { useLanguage } from '../../hooks/languageHook'
 
 const AboutText = () => {
 	const { language } = useLanguage()
