+++ src/components/Skills/Card.tsx
