+++ src/utils/imageToDataUrl.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent, Dispatch, SetStateAction } from 'react'
-import { IAllInfo } from '../components/UI/AddProject/AddNewProject'
+import { IAllInfo } from '../components/AddProject/AddNewProject'
 
 const imageToDataUrl = (
 	event: ChangeEvent<HTMLInputElement>,
